api/api/requirements.txt
asgiref==3.8.1
Django==4.2.11
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==24.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.11.0
mysqlclient == 2.2.0
